models/requirements-openvino.txt
openvino-dev[pytorch,onnx]
nncf==2.7.0
openai-whisper